+++ compose/src/jvmMain/kotlin/functions/clipboard.kt
@@ -1,6 +1,6 @@
 package functions
 
-import modules.singleton.QR
+import app.AppQR
 import java.awt.Desktop
 import java.awt.datatransfer.UnsupportedFlavorException
 import java.awt.image.BufferedImage
@@ -9,7 +9,7 @@ import java.net.URISyntaxException
 
 fun scanAndOpenQR(qrImage: BufferedImage) {
     try {
-        val scannedValue = QR.scan(qrImage)
+        val scannedValue = AppQR.scan(qrImage)
         val uri = URI(scannedValue)
         Desktop.getDesktop().browse(uri)
     } catch (e: UnsupportedFlavorException) {
