+++ compose/src/jvmMain/kotlin/views/Root.kt
@@ -2,28 +2,24 @@
 
 package views
 
-import androidx.compose.desktop.AppManager
 import androidx.compose.ui.window.MenuItem
 import components.common.NonSizeWindow
 import components.layout.AppTray
-import extentions.toTop
-import java.awt.Frame
+import app.AppWindows
 import kotlin.system.exitProcess
 
 fun Root() = NonSizeWindow {
     AppTray(
         MenuItem(
+            name = "Usage",
+            onClick = {
+                AppWindows.viewUsage()
+            }
+        ),
+        MenuItem(
             name = "QR Tool",
             onClick = {
-                val appFrame = AppManager.windows.firstOrNull()
-                if (appFrame == null) {
-                    Main()
-                } else {
-                    if (appFrame.window.state == Frame.ICONIFIED) {
-                        appFrame.window.state = Frame.NORMAL
-                    }
-                    appFrame.window.toTop()
-                }
+                AppWindows.viewQRScreenshot()
             }
         ),
         MenuItem(
