+++ server/src/main/resources/static/index.html
@@ -1,1 +1,1 @@
-<!DOCTYPE html><html lang="en"><head><meta charset="utf-8"><meta http-equiv="X-UA-Compatible" content="IE=edge"><meta name="viewport" content="width=device-width,initial-scale=1"><link rel="icon" href="/favicon.ico"><title>vue-template</title><link href="/js/about.5554677a.js" rel="prefetch"><link href="/css/app.7fce07d8.css" rel="preload" as="style"><link href="/js/app.89d355ff.js" rel="preload" as="script"><link href="/js/chunk-vendors.de7b539e.js" rel="preload" as="script"><link href="/css/app.7fce07d8.css" rel="stylesheet"></head><body><noscript><strong>We're sorry but vue-template doesn't work properly without JavaScript enabled. Please enable it to continue.</strong></noscript><div id="app"></div><script src="/js/chunk-vendors.de7b539e.js"></script><script src="/js/app.89d355ff.js"></script></body></html>
+<!DOCTYPE html><html lang="en"><head><meta charset="utf-8"><meta http-equiv="X-UA-Compatible" content="IE=edge"><meta name="viewport" content="width=device-width,initial-scale=1"><link rel="icon" href="/favicon.ico"><title>vue-template</title><link href="/js/about.5554677a.js" rel="prefetch"><link href="/css/app.7fce07d8.css" rel="preload" as="style"><link href="/js/app.da77aa0c.js" rel="preload" as="script"><link href="/js/chunk-vendors.de7b539e.js" rel="preload" as="script"><link href="/css/app.7fce07d8.css" rel="stylesheet"></head><body><noscript><strong>We're sorry but vue-template doesn't work properly without JavaScript enabled. Please enable it to continue.</strong></noscript><div id="app"></div><script src="/js/chunk-vendors.de7b539e.js"></script><script src="/js/app.da77aa0c.js"></script></body></html>
